requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow==10.0.1
lxml==4.9.3
urllib3==2.0.7
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3 